pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "malware-triage"
version = "0.1.0"
description = "Portable CLI malware triage and analysis tool with transparent confidence scoring"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pefile>=2023.2.7",
    "requests>=2.31.0",
    "rich>=13.7.0",
    "jinja2>=3.1.3",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
]

[project.optional-dependencies]
analysis = [
    "python-magic>=0.4.27",
    "py-tlsh>=4.12.0",
    "ppdeep>=20200000",
    "yara-python>=4.3.0",
    "oletools>=0.60.1",
    "pcodedmp>=1.2.6",
    "XLMMacroDeobfuscator>=0.2.7",
    "peepdf-3>=4.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]
dynamic = [
    "speakeasy-emulator>=1.5.9",
]

[project.scripts]
malware-triage = "main:cli"

[tool.setuptools]
py-modules = ["main"]
packages = [
    "core",
    "modules",
    "modules.static",
    "modules.enrichment",
    "modules.dynamic",
    "reporting",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
